requirements.txt
tinytag
music-tag
pytaglib
